requirements.txt
Django==2.1.7
django-enumfield==1.5
django-filter==2.1.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
djangorestframework-stubs==0.3.0
pkg-resources==0.0.0
pycodestyle==2.5.0
PyJWT==1.7.1
pytz==2018.9
